requirements.txt
# Google ADK
google-adk
google-genai

# RAG: vector store and embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.0

# Hugging Face datasets
datasets>=2.14.0
huggingface_hub>=0.19.0

# Optional: load RAGTruth from GitHub
requests>=2.28.0
